chore(templates): Move common things to angular-templates-runtime
