+++ packages/angular-templates-runtime/index.js
@@ -1,0 +1,4 @@
+AngularTemplates = {
+  scanner: HtmlScan.init,
+  handler: StaticHtmlTagHandler.init
+};
