+++ packages/angular-templates-runtime/ng-html-scanner.js
@@ -1,8 +1,3 @@
-scanHtmlForTags = function (options) {
-  const scan = new HtmlScan(options);
-  return scan.getTags();
-};
-
 /**
  * Scan an HTML file for top-level tags and extract their contents. Pass them to
  * a tag handler (an object with a handleTag method)
@@ -11,7 +6,7 @@ scanHtmlForTags = function (options) {
  * top-level tags, which are allowed to have attributes,
  * and ignores top-level HTML comments.
  */
-class HtmlScan {
+HtmlScan = class HtmlScan {
   /**
    * Initialize and run a scan of a single file
    * @param  {String} sourceName The filename, used in errors only
@@ -197,3 +192,8 @@ class HtmlScan {
     return this.tags;
   }
 }
+
+HtmlScan.init = function (options) {
+  const scan = new HtmlScan(options);
+  return scan.getTags();
+};
